feat: 添加日志和日志切面watchdog

+++ src/main/java/org/example/module/system/user/controller/SysUserController.java
@@ -1,13 +1,17 @@
 package org.example.module.system.user.controller;
 
-
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+
+import java.util.List;
+import java.util.Map;
+
 import org.example.common.base.BaseController;
 import org.example.common.domain.request.QueryRequest;
 import org.example.common.domain.response.Result;
+import org.example.common.util.ListUtil;
 import org.example.module.system.user.domain.entity.SysUser;
 import org.example.module.system.user.domain.request.LoginRequest;
 import org.example.module.system.user.domain.request.SysUserRequest;
@@ -32,10 +36,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-
 /**
  * <p>
  * 系统用户 前端控制器
@@ -50,7 +50,6 @@ public class SysUserController extends BaseController<SysUserServiceImpl, SysUse
 
     @Autowired
     private ISysUserRoleService sysUserRoleService;
-
 
     @PostMapping("/login")
     public Result login(@RequestBody @Validated LoginRequest request) {
@@ -80,8 +79,8 @@ public class SysUserController extends BaseController<SysUserServiceImpl, SysUse
         if (!StringUtils.isEmpty(query.getOrderField())) {
             queryWrapper.orderBy(true, query.getIsAsc(), query.getOrderField());
         }
-        IPage<SysUser> page = getBaseService().page(new Page<>(query.getPageIndex(), query.getPageSize()),
-                queryWrapper);
+        IPage<SysUser> page = getBaseService()
+            .page(new Page<>(query.getPageIndex(), query.getPageSize()), queryWrapper);
         return Result.success(page);
     }
 
@@ -102,12 +101,16 @@ public class SysUserController extends BaseController<SysUserServiceImpl, SysUse
     @PreAuthorize("hasAuthority('system:user:delete')")
     @DeleteMapping("/{ids}")
     public Result delete(@PathVariable Long[] ids) {
-        List<Long> idList = Arrays.asList(ids);
-        idList.forEach(id -> {
-            getBaseService().removeById(id);
-            sysUserRoleService.remove(new LambdaQueryWrapper<SysUserRole>()
-                    .eq(SysUserRole::getUserId, id));
-        });
+        // 去除admin
+        List<Long> idList = ListUtil.arr2List(ids);
+        ListUtil.removeIf(idList, (e) -> e.equals(1L));
+
+        if (ListUtil.isNotEmpty(idList)) {
+            idList.forEach(id -> {
+                getBaseService().removeById(id);
+                sysUserRoleService.remove(new LambdaQueryWrapper<SysUserRole>().eq(SysUserRole::getUserId, id));
+            });
+        }
         return Result.success();
     }
